Move ProcDetails to own file.
